More work to fix statistical combat

+++ src/com/carleson/battlecalc/Form.java
@@ -2,13 +2,13 @@ package com.carleson.battlecalc;
 
 public class Form {
 
-    public int OF=0;
-    public int DF=0;
-    public int AT=0;
-    public int size=0;
-    public int HE=0;
-    public int XP=0;
-    public int hits = 0;
+    public double OF=0;
+    public double DF=0;
+    public double AT=0;
+    public double size=0;
+    public double HE=0;
+    public double XP=0;
+    public double hits = 0;
     public String name = "";
